creating the routine function and adding the dead flag to the philosopher

+++ 42_philo/helper_functions/ft_atoi.c
@@ -25,9 +25,9 @@ static int	ft_isspace(const char c)
 
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
@@ -47,5 +47,5 @@ int	ft_atoi(const char *str)
 	}
 	if (result > 2147483647 || result < -2147483648)
 		return (0);
-	return (result * sign);
+	return (int)(result * sign);
 }
